Refactor the new structure that added with pr38
