+++ grapheneX.py
@@ -4,27 +4,13 @@
 from core.utils.argparser import parse_cli_args
 from core.utils.helpers import print_header
 from core.cli.shell import start_cli
-from core.web.server import run_server
-from core.utils.helpers import print_header, check_privileges, parser_host_port, logger
-from core.cli.shell import Shell
-from core.web import app
-
-DEBUG = True
+from core.web import run_server
 
 def main():
     args = parse_cli_args()
     print_header()
     if(args['web']):
         run_server(args)
-        host, port = parser_host_port(args['host_port'])
-        try:
-            logger.info(f'Starting Server {host}:{port}')
-            app.run(host=host, port=port, debug=DEBUG)
-        except:
-            logger.error('Invalid host & port address')
-            logger.info('Using default (host: 0.0.0.0, port: 8080)')
-            logger.info('Starting Server 0.0.0.0:8080')
-            app.run(host='0.0.0.0', port=8080, debug=DEBUG)
     else:
         start_cli()
